GL plugin: VSync option, some renaming, a little bit of MSAA preparations (not there yet)
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@2563 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Plugins/Plugin_VideoOGL/Src/Globals.cpp
@@ -63,7 +63,7 @@ bool LocalLogFile = true;
 #if LOGLEVEL > 0
 void __Log(const char *fmt, ...)
 {
-	int len = strlen(fmt);
+	size_t len = strlen(fmt);
 	if (!len)
 		return;
     char* Msg = (char*)alloca(len + 512);
